requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.3
stripe==7.11.0
sqlalchemy==2.0.27
alembic==1.13.1
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9
redis==5.0.1 